% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTopNAttractions.R
\name{getTopNAttractions}
\alias{getTopNAttractions}
\title{getTopNAttractions}
\usage{
getTopNAttractions(city, n)
}
\arguments{
\item{city}{A character vector of maximum length 1 (city requested)}

\item{n}{An integer value
(Number ot top spots to fectch for the chosen city)}
}
\value{
A dataframe containing all results fetched from Tripadvisor
and Twitter regarding the top attractions in the chosen city
(e.g. popularity level).
}
\description{
Provide a dataframe containing the data about the top N attractions
for a certain city requested by the user, combining hot spot search from
Tripadvisor and obtaining relevant data from Twitter (e.g. sentiments
about a certain place).
}
\details{
The function automatically obtains the most relevant places
for the selected city  based on a search on Tripdadvisor, then queries
Twitter about each hot spot (e.g. monument). \cr
Afterwards it runs a sentiment analysis on those selected N top attractions,
and also measures the popularitiy of the place
(using the count of the number of tweets mentioning the #place).  \cr  \cr
The function returns as a result a dataframe containing all the relevant
information about the top spots in that city combining in that sense also
live reviews from Twitter users, which can then be plotted interactively on a
Shiny map as well \code{\link{run_Shiny}}.
}
\examples{
\dontrun{getTopNAttractions(
city = "london",
n = 10
)
}

}
\references{
Orso, S., Molinari, R., Lee,
J., Guerrier, S., & Beckman, M. (2018).
An Introduction to Statistical Programming Methods with R.
Retrieved from \url{https://smac-group.github.io/ds/}
}
\seealso{
\code{\link{touristR} & \link{run_shiny}}
}
\author{
Ayrton Rua: \email{ayrton.gomesmartinsrua@unil.ch}

Maurizio Griffo: \email{maurizio.griffo@unil.ch}

Ali Karray: \email{mohamedali.karray@unil.ch}

Mohit Mehrotra: \email{mohit.mehrotra@unil.ch}

Youness Zarhloul: \email{youness.zarhloul@unil.ch}
}
