% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_keyword.R
\name{track_keyword}
\alias{track_keyword}
\title{track_keyword}
\usage{
track_keyword(keyword, number, sincetype, provideN)
}
\arguments{
\item{keyword}{A character vector of maximum length 4 (keywords requested)}

\item{number}{An integer value (Timeframe parameter)}

\item{sincetype}{A character (Timeframe parameter) can be (uniquely)
on of the following: either "days", "weeks", "months" or "years"}

\item{provideN}{An integer value
(Number ot tweets to fectch for each keyword)}
}
\value{
A timeseries scatter plot with the keyword count
for each requested keyword.

A dataframe containing all results fetched from Twitter (e.g. tweets)
based on package \code{\link{twitteR}}
}
\description{
Provide a timeline plot of keyword counts based on data obtained
from Twitter.
}
\details{
The function automatically obtains the latitude and longitude of
each requested place,then query the Twitter API to obtain the number of
tweets with the requested # keywords for the desired timeframe, and returns
finally at the end a \code{\link{ggplot2}} scatterplot,
allowing thus to analyze which places are the most popular
over a given time period. \cr
It should be noted in that sense that if the time frame selected by the user
is less than or equal to 2 weeks, the plot returned groups data by hour,
whereas if the user requests longer timeframes the result will be grouped by
days for better representation of the changes over time (short/long term). \cr \cr
Lastly, we advise the user to keep the parameter \code{provideN} short
to avoid long delay responses from Twitter API (e.g. setting provideN=100
would be preferable than 1000)
}
\examples{
\dontrun{track_keyword(
keyword = c("eiffel tower", "san francisco", "london"),
number = 2,
sincetype = "weeks",
provideN = 100
)
}

}
\references{
Orso, S., Molinari, R., Lee,
J., Guerrier, S., & Beckman, M. (2018).
An Introduction to Statistical Programming Methods with R.
Retrieved from \url{https://smac-group.github.io/ds/}
}
\seealso{
\code{\link{touristR} & \link{run_shiny}}
}
\author{
Ayrton Rua: \email{ayrton.gomesmartinsrua@unil.ch}

Maurizio Griffo: \email{maurizio.griffo@unil.ch}

Ali Karray: \email{mohamedali.karray@unil.ch}

Mohit Mehrotra: \email{mohit.mehrotra@unil.ch}

Youness Zarhloul: \email{youness.zarhloul@unil.ch}
}
