% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_twitter_hashtag.R
\name{track_twitter_hashtag}
\alias{track_twitter_hashtag}
\title{track_twitter_hashtag}
\usage{
track_twitter_hashtag(keyword, type, number, sincetype, provideN)
}
\arguments{
\item{keyword}{A character vector of maximum length 4 (keywords requested)}

\item{type}{A character vector of length 1 (type of keyword, can
be either a city/place/monument name in English, to be referred to as
"place", or if needed another type of keyword, please specify by "other").  \cr
Note: all keywords should be of the same types (e.g. all places' names).}

\item{number}{A positive integer number (timeframe parameter)}

\item{sincetype}{A character vector of length 1 (timeframe parameter),
which can be (uniquely) on of the following:
either "days", "weeks", "months" or "years"}

\item{provideN}{A positive integer number
(Number ot tweets to fetch per keywords)}
}
\value{
A timeseries scatterplot with the hashtag count on the y axis,
and on the x axis the timeframe.

A dataframe containing all the results fetched from Twitter
(e.g. tweets) based on package \code{\link{twitteR}}
}
\description{
Provide a scatterplot displaying hashtag counts overtime based on data obtained
from Twitter.
}
\details{
The function automatically obtains the latitude and longitude of
each requested place, then queries the Twitter API to obtain the number of
tweets with the requested # keywords for the desired timeframe, and returns
at the end a scatterplot based on the
aesthetics of \code{\link{ggplot2}},
allowing thus to analyze which places are the most popular
over a given time period. \cr
It should be noted in that sense that if the time frame selected by the user
is less than or equal to 2 weeks, the plot returned groups data by hour,
whereas if the user requests longer timeframes the result will be grouped by
days for better representation of the changes over time (short/long term). \cr \cr
Lastly, we advise the user to keep the parameter \code{provideN} low
to avoid long delay responses from Twitter API (e.g. setting provideN=100
would be preferable than 1000).
}
\examples{
\dontrun{track_keyword(
keyword = c("eiffel tower", "san francisco", "london"),
type="place",
number = 2,
sincetype = "weeks",
provideN = 100
)
}

}
\references{
Orso, S., Molinari, R., Lee,
J., Guerrier, S., & Beckman, M. (2018).
An Introduction to Statistical Programming Methods with R.
Retrieved from \url{https://smac-group.github.io/ds/}
}
\seealso{
\code{\link{touristR}} & \code{\link{run_shiny}}
}
\author{
Ayrton Rua: \email{ayrton.gomesmartinsrua@unil.ch}

Maurizio Griffo: \email{maurizio.griffo@unil.ch}

Ali Karray: \email{mohamedali.karray@unil.ch}

Mohit Mehrotra: \email{mohit.mehrotra@unil.ch}

Youness Zarhloul: \email{youness.zarhloul@unil.ch}
}
